+++ window.cpp
@@ -27,7 +27,7 @@ static void HandleImGui(const SDL_Event &event)
 	ImGuiIO &igio = ImGui::GetIO();
 	switch (event.type) {
 	case SDL_MOUSEMOTION:
-		igio.MousePos = ImVec2(event.motion.x, event.motion.y);
+		igio.MousePos = ImVec2((float)event.motion.x, (float)event.motion.y);
 		break;
 	case SDL_MOUSEBUTTONDOWN: {
 		int x = igMouseIndex(event.button.button);
